+++ src/components/Terminal/TerminalSettings/index.tsx
@@ -4,13 +4,12 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import {
     CollapsibleGroup,
     Dropdown,
     getPersistedTerminalSettings,
-    NumberInlineInput,
     persistTerminalSettings,
     selectedDevice,
     StateSelector,
@@ -22,7 +21,6 @@ import {
     getEchoOnShell,
     getLineEnding,
     getLineMode,
-    getScrollback,
     getSerialOptions,
     getSerialPort,
     LineEnding,
@@ -30,20 +28,15 @@ import {
     setEchoOnShell,
     setLineEnding,
     setLineMode,
-    setScrollback as setActiveScrollback,
 } from '../../../features/terminal/terminalSlice';
 import { convertToDropDownItems } from '../../../utils/dataConstructors';
 import ExportLog from './ExportLog';
-import HistoryFile from './HistoryFile';
 
 export default () => {
     const device = useSelector(selectedDevice);
     const serialPort = useSelector(getSerialPort);
     const lastModemOpenState = useRef(false);
     const selectedSerialport = useSelector(getSerialOptions).path;
-
-    const activeScrollback = useSelector(getScrollback);
-    const [scrollback, setScrollback] = useState(activeScrollback);
 
     const clearOnSend = useSelector(getClearOnSend);
     const lineEnding = useSelector(getLineEnding);
@@ -174,28 +167,9 @@ export default () => {
                         label="Device controls echo"
                     />
                 )}
-                <div
-                    title="Set the number of lines it is possible to scroll in the Terminal"
-                    className="tw-flex tw-justify-between tw-pt-1"
-                >
-                    Scrollback
-                    <NumberInlineInput
-                        value={scrollback}
-                        onChange={setScrollback}
-                        onChangeComplete={() => {
-                            if (scrollback !== activeScrollback) {
-                                dispatch(setActiveScrollback(scrollback));
-                            }
-                        }}
-                        range={{ min: 1, max: 2 ** 64 - 1 }}
-                    />
-                </div>
             </CollapsibleGroup>
             <CollapsibleGroup heading="Write to file">
                 <ExportLog />
-            </CollapsibleGroup>
-            <CollapsibleGroup heading="History File">
-                <HistoryFile />
             </CollapsibleGroup>
         </>
     );
