+++ src/features/settings/index.tsx
@@ -5,16 +5,13 @@
  */
 
 import React, { useEffect, useState } from 'react';
-import { ProgressBar } from 'react-bootstrap';
 import { useDispatch, useSelector } from 'react-redux';
 import {
-    Button,
     Card,
     FileLink,
     MasonryLayout,
     NumberInlineInput,
 } from '@nordicsemiconductor/pc-nrfconnect-shared';
-import { openFile } from '@nordicsemiconductor/pc-nrfconnect-shared/src/utils/open';
 
 import {
     getPathToHistoryFile,
@@ -29,6 +26,10 @@ import {
     getMaximumNumberOfLinesInHistory,
     getNumberOfLinesInHistory,
 } from '../history/historySlice';
+import {
+    getScrollback,
+    setScrollback as setActiveScrollback,
+} from '../terminal/terminalSlice';
 
 const HistorySettings = () => {
     const dispatch = useDispatch();
@@ -51,48 +52,83 @@ const HistorySettings = () => {
     }, [dispatch]);
 
     return (
-        <div className="tw-max-w-[512px]">
-            <Card title="History File">
-                <div className="tw-whitespace-nowrap tw-p-1">
-                    Current file usage: ({historyUsagePercentage.toFixed(0)}%)
-                </div>
+        <Card title="History File">
+            <div className="tw-flex tw-items-center tw-whitespace-nowrap tw-p-1">
+                <div className="tw-pr-2">Current file usage:</div>{' '}
+                {historyUsagePercentage.toFixed(0)}%
+            </div>
 
-                <div
-                    title="Set the maximum number of lines to store in the history file"
-                    className="tw-flex tw-items-center tw-justify-between tw-p-1"
-                >
-                    Max number of lines in history file{' '}
-                    <NumberInlineInput
-                        className="tw-p-0"
-                        value={newMaximumNumberOfLines}
-                        onChange={setNewMaximumNumberOfLines}
-                        onChangeComplete={size =>
-                            dispatch(setNewMaximumNumberOfLinesInHistory(size))
+            <div
+                title="Set the maximum number of lines to store in the history file"
+                className="tw-flex tw-items-center tw-p-1"
+            >
+                <div className="tw-pr-2">
+                    Max number of lines in history file:
+                </div>
+                <NumberInlineInput
+                    className="tw-p-0"
+                    value={newMaximumNumberOfLines}
+                    onChange={setNewMaximumNumberOfLines}
+                    onChangeComplete={size =>
+                        dispatch(setNewMaximumNumberOfLinesInHistory(size))
+                    }
+                    range={{
+                        min: MINIMUM_MAX_NUMER_OF_LINES,
+                        max: MAXIMUM_MAX_NUMBER_OF_LINES,
+                    }}
+                />
+            </div>
+
+            <div
+                title={pathToHistoryFile}
+                className="tw-flex tw-items-center tw-justify-between tw-gap-1 tw-overflow-hidden tw-whitespace-nowrap tw-p-1"
+            >
+                File Location:
+                <FileLink
+                    label={`${pathToHistoryFile}`}
+                    fileLocation={pathToHistoryFile}
+                />
+            </div>
+        </Card>
+    );
+};
+
+const TerminalScrollback = () => {
+    const dispatch = useDispatch();
+    const activeScrollback = useSelector(getScrollback);
+    const [scrollback, setScrollback] = useState(activeScrollback);
+
+    return (
+        <Card title="XTerm Scrollback">
+            <p>
+                The Terminal is only able to keep a limited amount of data in
+                its buffer. This amount is still adjustable using the scrollback
+                option below. This will increase/decrease the maximum number of
+                lines that can be kept in the buffer.
+            </p>
+            <div
+                title="Set the number of lines it is possible to scroll in the Terminal"
+                className="tw-flex tw-items-center tw-pt-1"
+            >
+                <div className="tw-pr-2">Scrollback:</div>
+                <NumberInlineInput
+                    value={scrollback}
+                    onChange={setScrollback}
+                    onChangeComplete={() => {
+                        if (scrollback !== activeScrollback) {
+                            dispatch(setActiveScrollback(scrollback));
                         }
-                        range={{
-                            min: MINIMUM_MAX_NUMER_OF_LINES,
-                            max: MAXIMUM_MAX_NUMBER_OF_LINES,
-                        }}
-                    />
-                </div>
-
-                <div
-                    title={pathToHistoryFile}
-                    className="tw-flex tw-items-center tw-justify-between tw-gap-1 tw-overflow-hidden tw-whitespace-nowrap tw-p-1"
-                >
-                    File Location:
-                    <FileLink
-                        label={`${pathToHistoryFile}`}
-                        fileLocation={pathToHistoryFile}
-                    />
-                </div>
-            </Card>
-        </div>
+                    }}
+                    range={{ min: 1, max: 2 ** 64 - 1 }}
+                />
+            </div>
+        </Card>
     );
 };
 
 export default () => (
-    <MasonryLayout minWidth={256}>
+    <MasonryLayout minWidth={256} className="tw-max-w-[1024px]">
+        <TerminalScrollback />
         <HistorySettings />
     </MasonryLayout>
 );
